+++ catalog/app/components/JsonEditor/constants.ts
@@ -2,7 +2,7 @@ import type { ErrorObject } from 'ajv'
 
 // import { JsonSchema } from 'utils/json-schema'
 
-import { JsonDictItem } from './State-with-types'
+import { JsonDictItem } from './State'
 
 // TODO: any JSON or EMPTY_VALUE
 export type JsonValue = $TSFixMe
